WE ARE ON AN EXPONENTIAL CURVE

+++ game/generator/world/world_gen.h
@@ -4,4 +4,4 @@
 #define ST_GEN_BASE_HEIGHT 0
 
 /// Returns the block height at a certain position
-i64 st_gen_get_block_height(i64 seed, i64 x, i64 z);
+f64 st_gen_get_block_height(i64 seed, i64 x, i64 z);
